[ALPS02968842] cmdq: Enable 64bit handle
[Detail]
Change handle and PA to 64 bit.

MTK-Commit-Id: 1ef08c7ba9005d5c6549262d12021e1d41518a5d

CR-Id: ALPS02968842
Feature: CMDQ

+++ drivers/misc/mediatek/cmdq/v3/cmdq_record.h
@@ -223,12 +223,12 @@ extern "C" {
  *     support only when secure OS enabled
  */
 	int32_t cmdq_op_write_reg_secure(struct cmdqRecStruct *handle, uint32_t addr,
-				   enum CMDQ_SEC_ADDR_METADATA_TYPE type, uint32_t baseHandle,
+				   enum CMDQ_SEC_ADDR_METADATA_TYPE type, uint64_t baseHandle,
 				   uint32_t offset, uint32_t size, uint32_t port);
 	int32_t cmdqRecWriteSecure(struct cmdqRecStruct *handle,
 				   uint32_t addr,
 				   enum CMDQ_SEC_ADDR_METADATA_TYPE type,
-				   uint32_t baseHandle,
+				   uint64_t baseHandle,
 				   uint32_t offset, uint32_t size, uint32_t port);
 
 /**
